Check column size on open from variables window
