+++ src/client/datascience/interactive-common/interactiveBase.ts
@@ -48,7 +48,7 @@ import { generateCellRangesFromDocument } from '../cellFactory';
 import { CellMatcher } from '../cellMatcher';
 import { addToUriList, translateKernelLanguageToMonaco } from '../common';
 import { Commands, Identifiers, Settings, Telemetry } from '../constants';
-import { ColumnWarningSize, IDataViewerFactory } from '../data-viewing/types';
+import { IDataViewerFactory } from '../data-viewing/types';
 import {
     IAddedSysInfo,
     ICopyCode,
@@ -101,6 +101,7 @@ import {
     WebViewViewChangeEventArgs
 } from '../types';
 import { WebviewPanelHost } from '../webviews/webviewPanelHost';
+import { DataViewerChecker } from './dataViewerChecker';
 import { InteractiveWindowMessageListener } from './interactiveWindowMessageListener';
 import { serializeLanguageConfiguration } from './serialization';
 
@@ -124,6 +125,7 @@ export abstract class InteractiveBase extends WebviewPanelHost<IInteractiveWindo
 
     protected abstract get notebookIdentity(): INotebookIdentity;
     protected fileInKernel: string | undefined;
+    protected dataViewerChecker: DataViewerChecker;
     private unfinishedCells: ICell[] = [];
     private restartingKernel: boolean = false;
     private perceivedJupyterStartupTelemetryCaptured: boolean = false;
@@ -205,6 +207,8 @@ export abstract class InteractiveBase extends WebviewPanelHost<IInteractiveWindo
                 l.onMessage(InteractiveWindowMessages.NotebookIdentity, this.notebookIdentity)
             );
         }, 0);
+
+        this.dataViewerChecker = new DataViewerChecker(configuration, applicationShell);
 
         // When a notebook provider first makes its connection check it to see if we should create a notebook
         this.disposables.push(
@@ -970,40 +974,9 @@ export abstract class InteractiveBase extends WebviewPanelHost<IInteractiveWindo
         }
     }
 
-    private async shouldAskForLargeData(): Promise<boolean> {
-        const settings = this.configuration.getSettings(this.owningResource);
-        return settings && settings.datascience && settings.datascience.askForLargeDataFrames === true;
-    }
-
-    private async disableAskForLargeData(): Promise<void> {
-        const settings = this.configuration.getSettings(this.owningResource);
-        if (settings && settings.datascience) {
-            settings.datascience.askForLargeDataFrames = false;
-            this.configuration
-                .updateSetting('dataScience.askForLargeDataFrames', false, undefined, ConfigurationTarget.Global)
-                .ignoreErrors();
-        }
-    }
-
-    private async checkColumnSize(columnSize: number): Promise<boolean> {
-        if (columnSize > ColumnWarningSize && (await this.shouldAskForLargeData())) {
-            const message = localize.DataScience.tooManyColumnsMessage();
-            const yes = localize.DataScience.tooManyColumnsYes();
-            const no = localize.DataScience.tooManyColumnsNo();
-            const dontAskAgain = localize.DataScience.tooManyColumnsDontAskAgain();
-
-            const result = await this.applicationShell.showWarningMessage(message, yes, no, dontAskAgain);
-            if (result === dontAskAgain) {
-                await this.disableAskForLargeData();
-            }
-            return result === yes;
-        }
-        return true;
-    }
-
     private async showDataViewer(request: IShowDataViewer): Promise<void> {
         try {
-            if (await this.checkColumnSize(request.columnSize)) {
+            if (await this.dataViewerChecker.checkColumnSize(request.columnSize, this.owningResource)) {
                 const jupyterVariableDataProvider = await this.jupyterVariableDataProviderFactory.create(
                     request.variable,
                     this._notebook!
